Move future async-colored files to _async
The old files are going to replaced with simple imports in the next
commit, and doing that at one fell swoop was confusing git blame.
Unfortunately, this does not help with `git diff` as of git 2.15.
